Create Master Sheet Capstone Project
This Master Sheet document is used as a record of individual progress in the Bangkit Academy 2024 capstone project. Each team member is expected to fill out this worksheet independently to record their activities and contributions during the checkpoint period. This Master Sheet is also the basis for individual assessment based on achievements at each checkpoint.